Add option to read entire file for letter frequency.
If user passes the option "--fast", then buildfreq.py will read the
entire input file into memory before performing operations on its
contents. This could result in the script using an amount of memory
that's twice the size of the input file. If this option is not passed,
then the input file will be read line-by-line.